Fix(web): allow upload button to work with tab key
To upload files, we use a hidden file input element inside a button. The input
got triggered by the associated label rendered in the button. The button itself
was doing nothing on a click.
This came with two issues. First of all, the button has a padding which, if
clicked, does not trigger the input as only the label does so. Second, the
button was doing nothing when selected with the keyboard. But if the user hit
`<tab>` another time, he was focusing an invisible element, the hidden file
input.

The solution for this is two folded. Clicking the button does now internally
trigger a click on the input element. Thereby the padding is no more an issue
and the keyboard interactions works too. Furthermore the hidden input element is
no more selectable to avoid confusion.

As the accessibility rules require is to use a label with the input element
inside, this put some additional challenges. Because if label got clicked, this
caused two open two file selection dialog windows as the button and the label
triggered the input element at the same time. The chosen solution sets the
button action to only trigger if the button itself gets clicked without its
content. This makes it working for the keyboard interaction and the padding
region. The label area is distinct to this and complements the button.

The problem with the mouse clicking through two elements is hard to discover
with unit tests that trigger events directly on DOM elements. We should consider
to re-introduce the `user-events` library for such scenarios.