Fix(web): allow to focus original document section
In terms of keyboard accessibility it was a problem that the original document
was not reachable with the `<tab>` key to scroll inside.
As this a "pure" `<div>` without anything special to the HTML specification,
there is no native detection as an intractable item by the browser. Thereby it
is necessary to manually set a `tabindex`. The challenge was to do this for an
element that gets generated by a dependency component (the TipTap editor).

RISDEV-263