Build and publish seperate frontend image
RISDEV-123
Separate steps for creating a standalone frontend image have now been added in the pipeline to deploy the backend and frontend in two different pods in the future. Additional steps in the pipeline:

- Build and publish frontend container image: here we run two new package scripts build:docker and publish:docker. A new image of our web folder is pushed to the registry at https://github.com/orgs/digitalservicebund/packages/container/package/ris-frontend.
- Sign the published front-end Docker image: this is where the image is signed
- Deploy a new frontend image: the new image is passed to our GitHub Action argocd-deploy https://github.com/digitalservicebund/github-actions/tree/main/argocd-deploy

In addition, the BACKEND_HOST variable in compose.yaml has been converted to a VITE_BACKEND_HOST variable, which is now defined in .env.development. This gives us access to this variable both in our local setup when the development server is running via compose setup, and in our production frontend image, which is built for production but for local testing via the ARG BUILD_MODE=development in the Dockerfile.prod also gets access to development variables. All API calls are thus correctly forwarded to the backend. In the long term, however, the BUILD_MODE argument should disappear from Dockerfile.prod, since it is only needed for local testing (which we actually have our compose setup for) and we only want to build the production image here. This also removed the proxy redirection in the site configuration.

To avoid a cross-origin error, all origins were allowed in the CorsGlobalConfiguration for the local profile. For the time of the conversion, the frontend is built in two directories (vite config build script) in parallel, in the backend and in the default web dist folder. In the future the frontend should no longer be built into the backend.