Feat(web): add new atomic TextInput component
To develop towards a standalone component library with atomic visual
components, we start to extract our styled inputs into separated
components. The first of them is the text input with some simple
styling.

We have decided to define a full input element interface for all all
input components. They should be usable just as normal input elements.
Should mean they can be mound to a `v-model` or using the classic `value` and
`input` event approach. Although this would be currently possible by
fall-through attributes, we explicitly define this interface. This got
done to provide the full palette of code features. From completion, type
hints to linting. Furthermore it is only a question of time when the
first input component can't use fall-through attributes.

Moreover it was necessary to extend the ESLint setup. The rule to ensure
that each label has an associated input element got configured to also
recognize custom input components.