Feat(web): extract composable for input models
As we intend to provide a mostly homogeneous interface to all custom
input components to make them usable like input elements, the shared
setup got extracted into a new composable. This also allows to simply
the input components tests to not re-test this basic functionality over
and over again.

Thereby this got extracted from the just added `TextInput` component.
And it gets used by the `FileInputButton` now too.
Note that it is not possible to define component properties or emitted
events from within a composable. Furthermore do the `define*` compiler
macros only support very basic interface usage. Should mean it is not
possible to re-use the interface of the composable to save more lines of
code which must be replicated in each component right now. But this is
technically just not possible right now. But at least it is type-safe.

This includes a bunch of small adjustments to align both input
components more (including the tests). It is a question to answer how to
further share "logic" between the components.