Feat(web): add visual InputField component
This component represents the repeating pattern to visually group
a label, icon and input element together with some specific styling. As
this group is primarily used within forms, the associated term `field`
is used.

While applying the usage of this new component for the document core
data form, some styling refactoring has been made. We use flexbox to
make styling simple and working more in general (e.g. gaps rather than
margins). Furthermore we use the unit `rem` rather than `px`.
While doing so, I recognized that the autofill styling could be moved
down to the input component themselves.

An important change was the adaption of the ESLint configuration. The
rules for the accessibility checks that each label has an associated
input element. This rule led already to "issues" when we started to use
our custom input components. With the new input field component we run
into new issues because we use a slot. In such cases where label and
input element are decoupled in code via `<slot />, `<components />` or
other techniques, this rule is not powerful enough to find the relation.
We still want to follow this accessibility rule, but also maintain full
code flexibility. Therefore we have decided to disable this check on the
linting layer. But in the accessibility end-to-end tests, this
label-input relation gets checked too. But here it is code independent
it checks the fully generated DOM in the web-browser.
Please note that we currently do not run the accessibility end-to-end
tests for all sub-pages and all scenarios. Though this is in the
pipeline. Nevertheless due to disabling the linting check, it can lead
to rule breaks until we catch up in the end-to-end tests.