Feat(web): add new ExpandableContent component
This components consists of a header and a body and follows to toggle
the visibility of the body. This is a requirements for upcoming page
designs. The goal is to have a very generic component that can be used
anywhere such functionality is required. Therefore it also does
basically no styling by itself and only provides some convenience
features in its interface.

Additionally, a new testing library as extension for the
`testing-library` "framework" has been added. It adds more matchers that
improve the readability of the tests, but also allows to test more high
level. As an example it provides an `isVisible` matcher which tests for
all possible ways an element can be visible or not visible without
having bad to read tests that check implementation details.