Chore(web): add linting rules for testing library
Recently we added `@testing-library/jest-dom` for better readable and
high-level unit tests. To support the usage of these matchers, an ESLint
plugin has been added with rules that check the recommended usage. All
broken rules could be automatically fixed.

This commit includes a small fix of a development dependency that made
it into the build dependencies. Unfortunately does `npm` not fail for
the unknown `install --dev` command and simply adds the package anyway.