Feat(web): add InputGroup component
In the front-end we have the repeating pattern and need to render
a bunch of input fields related to a data object. As a first step we
had introduced an approach that allows to define these fields in a JSON
and generate the input elements in a loop.
The former implementation was a bit clumsy and not re-useable, focusing
on the core data of documentation units only. The new component is more
generic and a bit smarter.

The new component solves the problem of a wrong tab order as it was
using two columns instead of rows. With the new approach it is possible
to define how much column it should display, but uses rows and flexbox
under the hood. This fixes the order and also addresses issues like
responsibility on smaller screens.
Moreover does the new component use a `v-model` approach to bind the
data to the fields.

So far, this input group works for text input fields only.

Overall the whole application (which is limited to the `DocUnitCoreData`
so far) made the surrounding code much simpler and shorter. Though it
had to change a little.
There is still some "compatibility" code to make the conversation step
by step. For example the format of the field descriptions.

For the tests, the generator approach has been introduced. The idea is
to have functions which generate objects of a certain type with random
data. As function parameter it receives an optional partial description
of the related object type. This allows to write tests where such object
instances are needed, but write them in a way so they keep readable and
understandable. Due to the function the individual test cases keep slim
and define only these properties of an object they actually care about
for this specific case.
In relation to this, some generators might be shared between multiple
test files. Therefore the `test-helper` directory got re-used. To make
it "importable" without relative paths, a new path/resolve alias got
added (the commonly used `~` prefix).
Please also note that this test helper file requires additional Talisman
configuration because the `"ab..z"` string gets recognized as a potential
secret.