Fix(web): resolve recursion problem in input group
In the past we assumed it would be a useful feature in the `InputGroup`
component to merge the given model data and the internal data of the
inputs. The root of this was that currently we have no good approach to
ensure on compile time that the input field definitions match the model
value.
The issue with this was that it copies by reference to merge both
objects. As result the reactive model value updates created an infinite
loop. Part of this issue is that input groups should/must work with deep
reactivity to allow them get nested.
The solution is to avoid the coping and rather "throw away" the local
input data if the parent updates the model value in a "breaking" way.
Future improvements might allow to detect such issues earlier. In
practice the end-to-end test cover all the functionality checks.

RISDEV-311