Feat(web): introduce typed input field approach
The `InputGroup` component allows to specify a list of inputs that
should be rendered and mound their model values to an object. So far
this only works for text inputs (using the `TextInput` component). The
goal is to make this more flexible and allow also other types of inputs.

To support this higher level goal, a generic `InputElement` component
has been added. The idea of this component is to act like the standard
HTML `<input />` component, but use our custom input components. Thereby
it receives the attribute `type` (as `<input />` does). This property
defines which actual input component will be rendered. In forwards the
model value bound to itself to the input component. Furthermore it can
receive any further attributes for the specific input which it will bind
to the component instance.
These two properties, the type and the input specific attributes are now
part of the definition of an input field. Thereby the `InputGroup` can
simply instantiate the `InputElement`.

To support this "generic" feature with more typing constraints, several
interfaces and types have been defined. As these types are part of the
higher level concept, they are part of the `domain`. This concept of
having objects with editable (or not) properties of different kind and
the rendering into forms, is independent of the actual visual frontend
implementation. It is not Vue specific and would still be valid and
working if we would switch to React or Angular.
The types are quite straight forward in an extending manner. One of the
first and strongest benefits is, that it allows us to type-check our
field definitions. These have been split into separate TypeScript files
to keep them more organized and, more importantly, being typed. The type
interference is so smart, once a field defines the requires `type`
property TypeScript will check the whole field definition if it is
compatible, has missing properties, correct property types etc.
Furthermore do this types used in the Vue components which implement
this higher level concept. The components do no more define these types
locally in a top-down-dependency manner. Rather we have a dependency
inversion to the domain now.

Unfortunately is Vue not (yet) powerful to support fully generic
components. And even to the degree it is possible, it is
a hack/workaround that leads to ugly code that is still not fully
generic. In result it is currently not possible to provide 100% type
safeness. It is not (yet) possible to tell the `InputElement` that if
`type` is of a specific enumeration entry, the `modelValue` must be of
a specific type. E.g. for a text type, the value is a string. But for
a checkbox a boolean. The same goes for the input related attributes.
Nevertheless the typing got restricted to some degree. Further
improvements planned as soon as possible. After all, even with the old
feature set only, the typing it stronger than before.

Please note that the current file-directory-structure under `domain/`
is still quite chaotic. But it was so too before. What the "domain"
actually is for the front-end, still has to be defined completely.
Therefore the most simple and flat approach has been chosen. Anyway,
an `index.ts` is used now to represent the `domain` more as a complete
module rather than individual parts. Other parts of the code base should
not know about the internal details.

After of all, this commit is too big and should have been split at least
into two. My apologizes for this mistake.

RISDEV-311