Feat(web): add component to repeat model component
While the commit title mind read confusing, the concept is very simple
but generic:
There are components that use the model value concept of Vue. The new
component allows to repeat such a component for a list of the same model
type. In short this means that per entry in the list, the component
creates an instance of this model component and binds the entry as model
value to it.
Furthermore this component provides the frame to add new entries and
remove them. Due to the reactivity based on the list, the visual parts
gets updated automatically. If a new entry gets added, the model
component gets another instance with this entry. It appears below/next
to the instance for the previous list entry. New entries can be added by
the parent, but also with a button. Therefore a template of a default
"empty" value must be provided to create this new list entry.
The same goes for removing entries again. All reactive and automatic.

This component tries to be as generic as possible and don't make any
assumptions about the component inside or the data itself. It just
implements the above concept.
Therefore the component allows to customize all parts of its visual
appearance with slots. The slot defaults are only there to ensure the
basic functionality, also for tests.
More over the component forwards any remaining properties to each
instance of the model component. This means that the parent can but any
number of additional properties on this component and they will be
copied to all these instances. With this approach the component provides
a mechanism that is open but closed.

To get a clearer image, this concept has emerged from the feature
request to have input groups that can be repeated. Means the user can
add the same group of input fields again and again. The first usage of
this mechanism will be for the "Rechtszug" of a documentation unit.
But after all this component is highly re-usable, good to test and
adaptable.

If anything about the needs of this component changes, it can be changed
at a single point and applies then everywhere. But it has a single
responsibility with very clear boundaries.