Fix(web): reduce duplicated code for fields
The SonarCloud quality gate has identified the input field definitions as too
much duplicated code. This wasn't an issue in the past because they were defined
in JSON. But going back to JSON isn't an option anymore because we have now more
complex types that can't be represented properly in JSON. It simply wouldn't
work with the `type` attribute, no matter if it is an enumeration or a
conjunction of specific strings.

To reduce the duplication for now small helper functions are used. They use
strict function parameter ordering to remove all the property names. This isn't
a nice solution at all. It is only acceptable because modern IDEs will display
the parameter names (and thereby the property names) as virtual text. We need to
find a better solution here. For now this should satisfy our pipeline (sounds
very bad).