Feat(web): add previous decisions to doc units
Documentation units have a new data property `previousDecisions`. This data
should be displayed and being editable.

This data set is special as it is the first one that is a list of repeating data
fields. Having this in mind, all the previous work in regards of input groups
and repeatable model components has been done. Therefore it was basically just a
accumulation of these components to apply it for this use-case now.

The special case is that this list can be empty. But in the UI the user should
always see a first empty entry. But this empty entry must not be saved back to
the backend unless the user actually inserts some values.

RISDEV-315