Fix(web): reduce minimal input field width
On screens with FullHD and larger font sizes, there is not enough space to
display the input fields in two columns. While the "merge" effect is desired
for even smaller screens, it should still be two columns on FullHD.

Currently the layout of the categories route is not too nice. We are still stuck
on a unflexibel grid system. As this system is not too easy to refactor right
now. Dropping the minimal input width is an alternative option.