chore(backend): add temporary vulnerability ignore
The CVE-2022-27664 is a new vulnerability that has been published
recently and affects our already running production release. It is not
related to a recent change in the project itself.
The root of this issue comes from the Docker image build utility tool
for the backend. It is written in Go and somehow leaks into the final
production image. Though we don't know (yet) why this leak happens, we
are pretty sure that there is no active Go code running in the final
image. Thereby it should not be possible to be used for an exploit.

The building tool has already a merged pull request to update the
vulnerable Go dependency that fixes the issue. Unfortunately there is no
new release of the builder yet.

As this issue is not affecting the running application itself, we
decided to ignore this CVE for now. **BUT** as soon as a new release of
the builder (that includes the fix) is made, we will upgrade and remove
the ignore again. Therefore we have established a bunch of reminders and
subscriptions. Moreover we are tagging a specific image hash of the
builder to (hopefully) get the update recognized by the depandabot too.

After all we need to investigate why the builder is leaking its
dependencies into the final image at all.
Finally we need to discuss how to handle such vulnerability issues in
future in general in the security guild of the company.