Chore(web): add linting rules for scoped styling
Using scoped styling helps to avoid uncontrolled side-effects and the hassle to
find out what is still used, where it comes from and if something can be removed.

To support the usage of scoped style, linting rules help us. These rules are
also capable of recognizing not used styling blocks.

The `TextButton` got slightly refactored to allow it being re-used by other
components too. Therefore it also allows to be used with a slot.

The `TextEditor` was and still is a mess in terms of styling for the component
dependencies it is using. We need to keep an eye on it.

RISDEV-287