Fix(web): redefine global styling like fonts
With the introduction of scoped styling the font settings got lost. The issue
was that we inject via preprocessor options the global styling into each
component. But the style of a component gets scoped now. Thereby the `App` style
can't fall down to the childs anymore. In fact we were re-defining the same
styles over and over again.

This has been resolved by introducing an actual global stylesheet that is
responsible to set the few global styles we have.