Chore(web): remove component style injection
Using a preprocessor feature, we used to inject a bunch of global variables and
styles into each component (basically a prelude). This lead to problems recently
as it is very hard to follow where these styling come from.

To solve the problem, the injection has been removed and components have to
manually import them. Thereby it is more clear where the data comes from. In
future this will improve by using `@use` instead of `@import` which is generally
recommended by SASS.

The REM functions have been removed as we want to use REM directly or rather
proper design tags soon. It was used at a very few places.

The strip unit functions have been removed as they were used nowhere anymore
(but was not detected of course).

The fonts are no more a separate file as they are used by the global style sheet
only.