Test(web): remove pending file upload check
When a file gets uploaded, the UI shows some information to the user
that the process is ongoing. This visual part gets checked in the
end-to-end tests. But because the upload is very often too fast, the
test fails as this message doesn't render at all.
Unfortunately we are missing the tools/approach right now to test this
feature properly. As it isn't actually critical, this single check gets
removed for now.

This commit includes some unrelated formatting of the StyleLint
configuration that was broken.