Chore(web): add postcss scrollbar plugin
We want to reduce (or rather completely remove) vendor specific style
prefixes. Right now we disable this linting rule as we still need them.
One case where the AutoPrefixer tool does not work is the scrollbar.
But another plugin can be used to follow the latest CSS specification
for scrollbars and provide polyfills.