Chore(web): remove vendor style prefixes
Having the AutoPrefixer plugin for PostCSS in place, we barely need any
vendor prefixes for our styles. All used rules have been checked with
the support of the prefixer.
Note that the scrollbar got fixed in a previous commit already.

This allows us to remove the ignore of the rule which disallows to use
vendor prefixes that are supported.