Chore(web): apply more healthy linting rules
The vast majority of the optional rules that have been enabled where
followed already anyways (e.g. using composition API with `<script
setup>`). These rules are only enabled to ensure that also in future we
do not start to derive from this standard. And if we do, we should
notice it and discuss in the team.

Some rules have been voted on if there was an option to decide on what
the team prefers.

Mostly all of the rule breaks could be automatically fixed. Only a very
few required manual work.

One rule that stands out is the more strict sorting of attributes and
classes in templates. This follows the same approach as for imports to
have overall unified code styles that is easy to read, get used to
across all developers.