Chore(backend): upgrade tiny builder with fix
A new version of the tiny builder has been released that includes an
updated version of the Go dependency that has been reported be
vulnerable.
With this update we could also remove the Trivy ignore file again as we
are clean again.

Please note that the version pinning of the builder is still not working
with the DependaBot yet.