Chore(web): adapt document wrapper grid
Remove the Vuetify grid components and use normal CSS flexbox.
This also introduces a shared padding for all wrapped components.

The slot of the wrapper got an attributed of classes. These classes must
be manually applied by all components that use the wrapper. There are
cases where it must be possible to partially overwrite the classes or
control to which exact element to apply them (e.g. future side by side
view).

Moreover got all headers adapted to the new font styling with a shared
class.