Test(web): fix TextEditor testing issues
The end-to-end tests targeting the `TextEditor` component are very
problematic and unstable. Thereby it is currently blocking our work and
causing a lot of issues.

The `TextEditor` has a bunch of buttons in a menu bar. Depending on the
size of the editor component, the menu adapts itself to collapse some
buttons into dropdown menus and other techniques. This means depending
on the surrounding layout and the screen size, the width of the editor
changes and thereby the state of the menu.
As result, it is a horrible idea to test this at a specific location on
the webpage in the end-to-end tests. If a padding or margin changes, the
tests start to fail as the menu looks different now.

Furthermore it is currently very hard to identify buttons in the DOM.
Thereby we use techniques with accessing the n-th child `div` and
similar. When a test fails (where it shouldn't are) it is very hard to
debug and fix and these tests are unreadable.

In relation to this issue, the whole `TextEditor` component gets
refactored heavily to make it better usable. It will be a generic
algorithm's that gets fed with a "configuration"/data In result, the
whole logic of the menu size behavior will be well testable separately
of the component.
In the next layer, the actual buttons to show is an integration test of
this logic with the data set of buttons to put into the menu. It must be
goal of the integration test to ensure all buttons get added to the menu
which are expected to exist. But the test must be independent on the
current state of the menu depending on the width.

Having the current state with the old component, it must still be an
"integration" test to test the existence of the buttons. Else we must
always struggle of permanently failing tests for unrelated changes where
everything is actually fine.
Therefore this has been added to the component tests and removed from
the end-to-end tests.
Note that finding a button in the current setup is not so nice. This
will improve with future aria labels and the new usage of font icons
instead of Vuetify icons. With these changes the current test case will
become better readable.

Also note that due to the above issues, it is currently very hard to
test the functionality of the buttons. In near future, this will be much
better and easy to do. As it is very hard to do now, easy soon and the
whole issue here is blocking the development heavily rather than it is
helping us, this small part will be untested for a limited amount of
time from now on.