Chore(web): adapt grid and fonts of categories
This removes the former margin around the bodies of the document unit category
sections.
It also adapts the heading style tags according to the latest design.

Moreover, the texts category no longer uses Vuetify grid layout components but
normal CSS flexbox.

Also the HTML element identifier used to navigate/scroll to the category section
are defined in the parent component now and are no longer the responsibility of
the children.