Chore(web): adapt grid for side navigation bar
Replace Vuetify grid components with normal CSS flexbox.

This does not refactor the component any further an rather copy-paste
the Tailwind style tags. This component should be improved soon.

Note that there is currently a color problem due to overlaps of Vuetify
and Tailwind. These will be resolved soon when Vuetify got removed from
the project. Till then, the borders in the menu will differ.