Replace tab-symbol with 4 spaces
TipTap converts &emsp; upon pasting to <0x2003> which get's displayed as wrong symbol in some text editors. In reasonable time I could not figure out how to make TipTap not convert this symbol. The &nbsp; it leaves as it is. I also tried making (typography) extensions with input- and pasting-rules for TipTap, but it didn't work.
RISDEV-455