Docs(norms): shift over norms documentation
Since the former norms repository has been deleted, we have to shift
over some files. Here we add the already created documentation for the
API and domain model. Both are still in a very early stage and are just
important to get the documentation topic ongoing from the beginning.

The documentation of the documentation will be added later (e.g. how to
generate stuff from the sources).