Chore(backend): move application class up
Because we have now the two child packages `caselaw` and `norms` which
belong both to the Spring application, the `Application` class must move
up to the `ris` package.

Additionally the workaround for Jacoco to ignore the `Application` class
on the test-coverage has been removed, as this is gets handled by
SonarQube now.