Chore(pipeline): always use latest cosign version
The pipeline got broken due to a Cosign update we are not compatible
with. So far we get updates for the cosign-installer GitHub action, but
not the version of Cosign that gets installed. This is fixed in the
action configuration. By removing this optional action input we always
get the latest one. As Cosign is still somewhat experimental and there
are also breaking changes in minor version updates, this seems to be the
best solution. Better would be to also have dependabot to work for this
version too.