Test(backend): apply scope to archunit tests
As we will now have the two packages `caselaw` and `norms` which are
kinda independent in their architecture, also the related architecture
fitness tests must be scoped to the package.
While the test class was already part of the `caselaw` package, it was
operating on the `ris` package.
It is planned to have another test class for the `norms` architecture
fitness.