Chore(hooks): move secret check to pre-commit
So far the check for secrets runs on pre-push. This leads to some
trouble when working with files that have a checksum. These files (e.g.
`build.gradle.kts` or `README.md`) have false positives and are accepted
with a specific checksum. Adjusting these files always creates a new
checksum and needs to be verified again. Finally the new checksum must
be put into the `.talismanrc` and added to the commit.
If the check runs after all commits are bundled, it is cumbersome to
either interactively rebase the local history to integrate this change,
or append a "stupid" fix commit.
To circumvent this problem, we now check secrets pre-commit. The check
is fast enough to not be annoying on every commit.