Feat(backend): add kotlin setup and norms domain
The first Kotlin package for the norms has been added. The domain
includes only some entities and values so far. As the domain is not yet
much rich, the tests are very basic.

This also includes the adjustments to the Gradle build to compile
Kotlin. It also extends Spottless with KtLint.