Feat(backend): add load norm use-case
This adds the norms application package with a first use-case as input
port. This is a command use-case and gets implemented by a service. The
output port is specified as well.
A unit test for the service mocks the output port with the new `mockk`
dependency which provides better mocking for Kotlin than `mockito` does.