Feat(backend): add first norms persistence adapter
This adapter is like a "dummy" that work only in-memory. A proper
database adapter that connects to the Postgres database will be added
soon.

Having this in place, all ouput adapter interfaces are implemented and
the dependencies of the use case services are satisfied. Thereby they
can be loaded into the Spring application context. Though there is no
controller for them yet.
This also required (for now) to allow the application package to depend
on Spring annotations.