Chore(backend): extend norms package directories
Though the official guideline is to start the source folder with the
base package that all packages from here share. This removes the need
for these insane long chain of "empty" folders just to mirror the
package path.
Anyway we have issue with our quality gate at SonarCloud that can't find
the coverage of the Kotlin files in the Jacoco report. The current
assumption is that the package versus file path might be an issue. To
verify this assumption it is necessary to actually commit and push this
to the remote repository.
If this is not the issue, we will revert it. If it is the issue, we take
it as temporarily fix and hopefully find a way to make it work again
with shorter paths.