Feat(backend): add query norms controller
Another output adapter that exposes the load and list norm(s) use cases.
As they use the same response schema and thereby need the same mappers,
these use cases are kept together for now. But it is very likely that
this gets split soon. Already the tests are too long and big and that
before the features are even complete.