Docs(norms): update domain model for UUID usage
We changed from a custom domain object value to represent the `guid`
properties of our entities to the global concept of a UUID. In the code
we refer to the Java standard library.

Usually the domain model documentation should be changed first as the
result of a team agreement. This has been forgotten. So we need to
define a process for the future how to keep documentation and code
synchronized.