Rework file upload component
RISDEV-404

Adapted the file input component to new design. To keep the error handling consistent, make the code more comprehensible and avoid code duplication, we decided to only display upload errors in the error modal and not on drag. The in-drag error messages were removed, as they were not up to date and not consistent with the error handling in the file service. As the File Input now is not triggered by an actual Button anymore, the component was renamed.