Feat(backend): allow to update norm long title
The long title of a norm must be able to be updated by the documentary.
So far there is no additional business logic related to this update and
it can be applied directly on the entity property. Therefore it is
feasible to simply make this property a variable. In future this will
become a setter function that includes the additional checks and logic.