Test(frontend): add tests for http client
To enable us to refactor the http client module, it is necessary (or
rather extremely helpful) to first cover it with tests.

Because the http client is a thin wrapper for dependency inversion to
the Axios library, pure unit tests do not make too much sense. Rather an
integration test using the library is more fruitful.
Unfortunately this is not easily possible as it heavily depends on in
which environment the tests run. For example does the Sinon fake server
not work here, as it mocks the web-browser environment.
The best option that is widely adopted is the `axios-mock-adapter`
package. It isn't optimal, but quite good for our needs.