Refactor(frontend): separate navbar and toggle
As they `SideToggle` component and soon also the `NavbarSide` components
are meant to be generically re-usable, they got separated from each
other. It is the responsibility of the component user to orchestrate
them together. Thereby the `NavbarSide` becomes much more simple and
easy to use.

Note that this also adds some default values for the `SideToggle`
properties.