Refactor(frontend): use enum for side toggle
To avoid using unsafe strings for the control flow, the `SideToggle`
property for the direction has changed to and enumeration. Thereby we
much stronger type safety in the using components as well as the
`SideToggle` component itself.
The naming of the property has slightly changed to be more expressive.
So the user of the component gets an immediate image of what the option
means.
Furthermore some styling got simplified and unified without copy-paste.