Refactor(frontend): add toggle query composable
This compose makes the following feature re-usable where a toggle value
is synchronized with a query parameter of the route.
This concept is used at two places so far. The side navbar and original
file side panel. Both for case-law. And as we will need this more often
in future, the composable is a nice approach.

Please note that the `DocumentUnitCategories` and
`OriginalFileSidePanel` components are still quite messy and have been
only adopted to somehow work again. Though it has not become worse than
before.