Refactor(frontend): allow to re-use navbar side
The `NavbarSide` component was so far a highly caselaw focused
component. It also heavily mixed data and logic with a lot of duplicated
code.
To make the component much cleaner and also re-usable, the component can
be configured now. It allows to define a list of menu items which can
have child items again. So far only two levels of nesting are allowed.
True recursion might be added when it becomes necessary, but is not
worth the effort for now.
Because the menu item configuration is a more reactive thing for caselaw
as it requires to be fed with the current document number, as well as
the current route query, this configuration has been defined as
a composable function. Thereby the `DocumentUnitWrapper` can now compose
all the components and data together as necessary.
This also changes how the tests work and test for more features than
before with a higher coverage.

Note that while refacoring the `NavbarSide` component, also the styling
got touched. It is much more aligned with the original design now.
Though it might need a final touch in cooperation with a designer.